feat(tab-select): graduation from experimental to core
now that we have been using this component for a bit,
we feel confortable enough to move it to core
