+++ src/app/components/components/components.module.ts
@@ -25,6 +25,7 @@ import { AnimationsComponent } from './animations/animations.component';
 import { DataTableDemoComponent } from './data-table/data-table.component';
 import { PagingDemoComponent } from './paging/paging.component';
 import { SearchDemoComponent } from './search/search.component';
+import { TabSelectDemoComponent } from './tab-select/tab-select.component';
 import { DynamicFormsDemoComponent } from './dynamic-forms/dynamic-forms.component';
 import { CodeEditorDemoComponent } from './code-editor/code-editor.component';
 import { TextEditorDemoComponent } from './text-editor/text-editor.component';
@@ -58,7 +59,7 @@ import { MatDividerModule } from '@angular/material/divider';
 import { CovalentCommonModule, CovalentLayoutModule, CovalentMediaModule, CovalentExpansionPanelModule, CovalentFileModule,
          CovalentStepsModule, CovalentLoadingModule, CovalentDialogsModule, CovalentSearchModule, CovalentPagingModule,
          CovalentNotificationsModule, CovalentMenuModule, CovalentChipsModule, CovalentDataTableModule, CovalentJsonFormatterModule,
-         CovalentMessageModule, CovalentVirtualScrollModule } from '../../../platform/core';
+         CovalentMessageModule, CovalentVirtualScrollModule, CovalentTabSelectModule } from '../../../platform/core';
 import { CovalentHighlightModule } from '../../../platform/highlight';
 import { CovalentMarkdownModule } from '../../../platform/markdown';
 import { CovalentDynamicFormsModule } from '../../../platform/dynamic-forms';
@@ -99,6 +100,7 @@ import { TdTestDynamicComponent } from './dynamic-forms/dynamic-forms.component'
     TextEditorDemoComponent,
     NotificationsDemoComponent,
     VirtualScrollDemoComponent,
+    TabSelectDemoComponent,
     // External Dependencies
     NgxTranslateDemoComponent,
     TdTestDynamicComponent,
@@ -152,6 +154,7 @@ import { TdTestDynamicComponent } from './dynamic-forms/dynamic-forms.component'
     CovalentCodeEditorModule,
     CovalentTextEditorModule,
     DocumentationToolsModule,
+    CovalentTabSelectModule,
     TranslateModule,
     componentsRoutes,
     ToolbarModule,
