+++ src/platform/core/tab-select/package.json
