+++ src/platform/core/tab-select/public-api.ts
