+++ src/platform/core/tab-select/tab-option.component.ts
