+++ src/platform/core/tab-select/tab-select.component.spec.ts
