+++ src/platform/core/tab-select/tab-select.module.ts
