+++ src/platform/experimental/tab-select/index.ts
@@ -1,1 +1,1 @@
-export * from './public-api';
+export * from '@covalent/core/tab-select';
