+++ src/test-bed/main/main.component.html
@@ -5,6 +5,3 @@
 <div class="pad">
   <a [routerLink]="'/breadcrumbs'">Breadcrumbs</a>
 </div>
-<div class="pad">
-  <a [routerLink]="'/tabselect'">Tab Select</a>
-</div>
