--- src/test-bed/sandbox/tab-select/tab-select.component.html
@@ -1,70 +1,0 @@
-<div class="pad">
-  <h2>Tab Select</h2>
-  <h3>Default with accent color with a disabled tab</h3>
-  <td-tab-select #tabSelect color="accent">
-    <td-tab-option value="test">
-      MY LABEL
-    </td-tab-option>
-    <td-tab-option value="test2">
-      MY LABEL 2
-    </td-tab-option>
-    <td-tab-option value="test3">
-      MY LABEL 3
-    </td-tab-option>
-  </td-tab-select>
-  <p class="tc-grey-600">value: {{tabSelect.value}}</p>
-
-  <h3>Value usage with primary backgroundColor</h3>
-  <td-tab-select [(value)]="value" backgroundColor="primary">
-    <td-tab-option value="test">
-      MY LABEL
-    </td-tab-option>
-    <td-tab-option value="test2">
-      MY LABEL 2
-    </td-tab-option>
-    <td-tab-option value="test3">
-      MY LABEL 3
-    </td-tab-option>
-  </td-tab-select>
-  <p class="tc-grey-600">value: {{value}}</p>
-
-  <h3>Forms Usage</h3>
-  <td-tab-select [(ngModel)]="formValue">
-    <td-tab-option value="test">
-      MY LABEL
-    </td-tab-option>
-    <td-tab-option value="test2">
-      MY LABEL 2
-    </td-tab-option>
-    <td-tab-option value="test3">
-      MY LABEL 3
-    </td-tab-option>
-  </td-tab-select>
-  <p class="tc-grey-600">value: {{formValue}}</p>
-
-  <h3>Disabled tabs</h3>
-  <td-tab-select disabled>
-    <td-tab-option>
-      MY LABEL
-    </td-tab-option>
-    <td-tab-option>
-      MY LABEL 2
-    </td-tab-option>
-    <td-tab-option>
-      MY LABEL 3
-    </td-tab-option>
-  </td-tab-select>
-
-  <h3>Stretched and disabled ripple on tabs</h3>
-  <td-tab-select stretchTabs disableRipple>
-    <td-tab-option>
-      MY LABEL
-    </td-tab-option>
-    <td-tab-option>
-      MY LABEL 2
-    </td-tab-option>
-    <td-tab-option>
-      MY LABEL 3
-    </td-tab-option>
-  </td-tab-select>
-</div>
