--- src/test-bed/sandbox/tab-select/tab-select.component.ts
@@ -1,13 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'tab-select-demo',
-  styleUrls: [ './tab-select.component.scss' ],
-  templateUrl: './tab-select.component.html',
-})
-export class TabSelectDemoComponent {
-
-  value: string = 'test3';
-  formValue: string = 'test2';
-
-}
