+++ src/test-bed/test-bed.module.ts
@@ -7,19 +7,16 @@ import { MatDividerModule } from '@angular/material/divider';
 import { MatToolbarModule } from '@angular/material/toolbar';
 
 import { CovalentBreadcrumbsModule } from '../platform/experimental/breadcrumbs/breadcrumbs.module';
-import { CovalentTabSelectModule } from '../platform/experimental/tab-select';
 
 import { TestBedComponent } from './test-bed/test-bed.component';
 import { MainComponent } from './main/main.component';
 import { BreadcrumbDemoComponent } from './sandbox/breadcrumbs/breadcrumbs.component';
-import { TabSelectDemoComponent } from './sandbox/tab-select/tab-select.component';
 import { appRoutes, appRoutingProviders } from './test-bed.routes';
 
 @NgModule({
   declarations: [
     TestBedComponent,
     BreadcrumbDemoComponent,
-    TabSelectDemoComponent,
     MainComponent,
   ],
   imports: [
@@ -32,7 +29,6 @@ import { appRoutes, appRoutingProviders } from './test-bed.routes';
     appRoutes,
     /** Experimental Modules */
     CovalentBreadcrumbsModule,
-    CovalentTabSelectModule,
   ], // modules needed to run this module
   providers: [
     appRoutingProviders,
