+++ src/test-bed/test-bed.routes.ts
@@ -1,7 +1,6 @@
 import { Routes, RouterModule } from '@angular/router';
 
 import { BreadcrumbDemoComponent } from './sandbox/breadcrumbs/breadcrumbs.component';
-import { TabSelectDemoComponent } from './sandbox/tab-select/tab-select.component';
 import { MainComponent } from './main/main.component';
 
 const routes: Routes = [
@@ -12,10 +11,6 @@ const routes: Routes = [
   {
     component: BreadcrumbDemoComponent,
     path: 'breadcrumbs',
-  },
-  {
-    component: TabSelectDemoComponent,
-    path: 'tabselect',
   },
 ];
